Rename helper functions and test to use camelCase
