+++ solidjs-tailwind/src/helper/searchFunctionality.js
