+++ solidjs-tailwind/src/helper/typeFilterFunction.js
